Add files via upload
When executed, this program will ask for the number of drawers and the type of wood for a desk.  It will use a function to calculate the costs of all the drawers, another one to calculate the cost of the desk based on the wood.  Finally, a third function will add up the total cost based on the wood type and drawer number.  Total cost, type of wood and the extra cost of drawers are displayed to the user.